pyproject.toml
[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[project]
name = "ipt_survey"
version = "2025.0.2"
authors = [
    {name = "Oleh Shkalikov", email = "Shkalikov.Oleh@gmail.com"},
]
description = "Automation of the IPT survey"
requires-python = ">=3.13"
license = {file = "LICENSE"}
readme = "README.md"
dependencies = [
    "google-api-python-client>=2.179.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pyarrow>=21.0.0",
    "python-telegram-bot[job-queue,rate-limiter]>=22.3",
    "tqdm>=4.67.1",
]

[tool.uv.build-backend]
module-name = "src"
module-root = ""
